% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorselect.R
\name{get_mirror}
\alias{get_mirror}
\title{get_mirror}
\usage{
get_mirror(repo = "CRAN")
}
\arguments{
\item{repo}{one of 'CRAN' or 'BioC'}
}
\value{
a vector of mirror urls
}
\description{
extract mirror lists from https://cran.r-project.org/mirrors.html
}
\examples{
head(get_mirror())
}
\author{
Guangchuang Yu
}
