% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorselect.R
\name{mirrorselect}
\alias{mirrorselect}
\title{mirrorselect}
\usage{
mirrorselect(mirrors)
}
\arguments{
\item{mirrors}{a vector of CRAN mirrors}
}
\value{
data frame with a column of mirror and second column of speed
}
\description{
test download speed of CRAN mirrors
by recording download time for mirror/src/base/COPYING
}
\author{
Guangchuang Yu
}
