% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirrorselect.R
\name{get_mirror}
\alias{get_mirror}
\title{get_mirror}
\usage{
get_mirror()
}
\value{
a vector of mirror urls
}
\description{
extract mirror lists from https://cran.r-project.org/mirrors.html
}
\author{
Guangchuang Yu
}
